+++ bitcoin/src/hash_types.rs
@@ -8,28 +8,6 @@
 //! hash).
 //!
 
-#[rustfmt::skip]
-macro_rules! impl_asref_push_bytes {
-    ($($hashtype:ident),*) => {
-        $(
-            impl AsRef<$crate::blockdata::script::PushBytes> for $hashtype {
-                fn as_ref(&self) -> &$crate::blockdata::script::PushBytes {
-                    use $crate::hashes::Hash;
-                    self.as_byte_array().into()
-                }
-            }
-
-            impl From<$hashtype> for $crate::blockdata::script::PushBytesBuf {
-                fn from(hash: $hashtype) -> Self {
-                    use $crate::hashes::Hash;
-                    hash.as_byte_array().into()
-                }
-            }
-        )*
-    };
-}
-pub(crate) use impl_asref_push_bytes;
-
 #[deprecated(since = "0.0.0-NEXT-RELEASE", note = "use crate::T instead")]
 pub use crate::{
     BlockHash, FilterHash, FilterHeader, TxMerkleNode, Txid, WitnessCommitment, WitnessMerkleNode,
